ajustando contenido

+++ app/Http/Controllers/PerfilController.php
@@ -20,6 +20,19 @@ class PerfilController extends Controller
     {
         $perfil = Auth::user()->perfil;
 
+        //Si por algún motivo no existiera lo creamos.
+        if(!$perfil){
+            $perfil = Auth()->user()->perfil()->create([
+                'tipo_documento',
+                'documento',
+                'movil',
+                'logo',
+                'descripcion',
+                'web',
+                'redes_sociales'
+            ]);
+        }
+
         return response([
             "error"=>false,
             "message"=>"Perfil obtenido correctamente",
@@ -32,7 +45,7 @@ class PerfilController extends Controller
      */
     public function update(Request $request)
     {
-        $perfil = Auth::user()->perfil;
+        $perfil = Auth::user()->perfil()->firstOrCreate([]);
 
         $validated = $request->validate([
             'tipo_documento' => 'nullable|in:nif,cif,nie',
@@ -42,13 +55,13 @@ class PerfilController extends Controller
             'web' => 'nullable|string|max:255',
             'redes_sociales' => 'nullable|array'
         ]);
-
+        //Aquí ya no tocamos el logo,solo los datos de texto.
         $perfil->update($validated);
 
         return response([
             "error" => false,
             "message" => "Perfil actualizado correctamente.",
-            "data" => $perfil
+            "data" => $perfil //Ahora incluirá 'logo_url' automáticamente.
         ], 200);
     }
 
@@ -56,24 +69,26 @@ class PerfilController extends Controller
      * Subir o actualizar el logo del usuario.
      */
     public function uploadLogo(Request $request){
-        $perfil = Auth::user()->perfil;
+        //Busca el perfil del usuario, si no existe, lo crea.
+        $perfil = Auth::user()->perfil()->firstorCreate([]);
         $request->validate([
             'logo'=>'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
         ]);
-        //Borrar logo anterior si existe.
-        if($perfil->logo){
+        //Borrar logo anterior si existe(y el archivo si existe).
+        if($perfil->logo && Storage::disk('public')->exists($perfil->logo)){
             Storage::disk('public')->delete($perfil->logo);
         }
 
-        //Guardar nuevo logo.
+        // El directorio 'logos' se creará solo dentro de storage/app/public/
         $path = $request->file('logo')->store('logos','public');
+
         $perfil->update([
             'logo' => $path
         ]);
         return response([
             "error" => false,
             "message" => "Logo actualizado correctamente.",
-            "data" => $perfil
+            "data" => $perfil //Incluirá logo_url al añadir el accessor.
         ],200);
     }
